requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.5
flask==3.0.1
geopandas==0.13.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pymongo==4.1.0
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
